changed something

+++ backend/src/main/java/at/htlleonding/dto/FetchedGroupDto.java
@@ -4,5 +4,5 @@ import at.htlleonding.model.Player;
 
 import java.util.List;
 
-public record FetchedGroupDto(Long id, String name, String link, List<Player> players, Integer count, boolean answered) {
+public record FetchedGroupDto(Long id, String name, String link, List<Player> players, Long count, boolean answered) {
 }
